backend/requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.39.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fsspec==2025.3.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kafka-python==2.1.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.0.1
simple-websocket==1.1.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
